requirements.txt
pytz
Flask==2.0.2
loguru==0.5.3
pytest==5.3.2
gunicorn==20.1.0
jsonschema[format, format_nongpl]==3.2.0
google-cloud-datastore==2.4.0
